Move contrib/hooks/python/hooks to contrib/hooks/python/training.
Experimental hooks will move to tensorflow/python/training if they prove
useful, so move files to reflect this.

- Fixups after move.
- Add hooks to cmake build.
